requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anthropic==0.35.0
    # via -r requirements.in
anyio==4.5.2
    # via
    #   anthropic
    #   httpx
    #   openai
appnope==0.1.4
    # via
    #   ipykernel
    #   ipython
asttokens==2.4.1
    # via stack-data
attrs==24.2.0
    # via
    #   cattrs
    #   e2b
    #   e2b-code-interpreter
    #   jsii
aws-cdk-asset-awscli-v1==2.2.214
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.3
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk-aws-apprunner-alpha==2.171.0a0
    # via -r requirements.in
aws-cdk-aws-lambda-go-alpha==2.171.0a0
    # via -r requirements.in
aws-cdk-aws-lambda-python-alpha==2.171.0a0
    # via -r requirements.in
aws-cdk-cloud-assembly-schema==38.0.1
    # via aws-cdk-lib
aws-cdk-lib==2.171.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-apprunner-alpha
    #   aws-cdk-aws-lambda-go-alpha
    #   aws-cdk-aws-lambda-python-alpha
aws-lambda-powertools==3.3.0
    # via -r requirements.in
backcall==0.2.0
    # via ipython
boto3==1.35.81
    # via -r requirements.in
botocore==1.35.81
    # via
    #   boto3
    #   s3transfer
cattrs==24.1.2
    # via jsii
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
comm==0.2.2
    # via ipykernel
constructs==10.4.2
    # via
    #   -r requirements.in
    #   aws-cdk-aws-apprunner-alpha
    #   aws-cdk-aws-lambda-go-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
debugpy==1.8.9
    # via ipykernel
decorator==5.1.1
    # via ipython
distro==1.9.0
    # via
    #   anthropic
    #   openai
e2b==1.0.4
    # via e2b-code-interpreter
e2b-code-interpreter==1.0.0
    # via -r requirements.in
executing==2.1.0
    # via stack-data
filelock==3.16.1
    # via huggingface-hub
fsspec==2024.10.0
    # via huggingface-hub
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via
    #   e2b
    #   httpx
httpx==0.27.2
    # via
    #   anthropic
    #   e2b
    #   e2b-code-interpreter
    #   openai
huggingface-hub==0.26.2
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.4.5
    # via jsii
ipykernel==6.29.5
    # via -r requirements.in
ipython==8.12.3
    # via ipykernel
jedi==0.19.2
    # via ipython
jiter==0.7.1
    # via
    #   anthropic
    #   openai
jmespath==1.0.1
    # via
    #   aws-lambda-powertools
    #   boto3
    #   botocore
jsii==1.105.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-apprunner-alpha
    #   aws-cdk-aws-lambda-go-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
nest-asyncio==1.6.0
    # via ipykernel
openai==1.55.1
    # via -r requirements.in
packaging==24.2
    # via
    #   e2b
    #   huggingface-hub
    #   ipykernel
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==4.3.6
    # via jupyter-core
prompt-toolkit==3.0.48
    # via ipython
protobuf==5.28.3
    # via e2b
psutil==6.1.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-apprunner-alpha
    #   aws-cdk-aws-lambda-go-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
pure-eval==0.2.3
    # via stack-data
pydantic==2.10.1
    # via
    #   anthropic
    #   openai
pydantic-core==2.27.1
    # via pydantic
pygments==2.18.0
    # via ipython
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   e2b
    #   jsii
    #   jupyter-client
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via huggingface-hub
pyzmq==26.2.0
    # via
    #   ipykernel
    #   jupyter-client
requests==2.32.3
    # via huggingface-hub
s3transfer==0.10.4
    # via boto3
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
stack-data==0.6.3
    # via ipython
tokenizers==0.20.3
    # via anthropic
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-apprunner-alpha
    #   aws-cdk-aws-lambda-go-alpha
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.12.2
    # via
    #   anthropic
    #   aws-lambda-powertools
    #   e2b
    #   huggingface-hub
    #   jsii
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==1.26.20
    # via
    #   botocore
    #   requests
wcwidth==0.2.13
    # via prompt-toolkit
